fix #1608

+++ src/Snap.Hutao/Snap.Hutao/ViewModel/Cultivation/CultivationViewModel.cs
@@ -1,6 +1,7 @@
 ﻿// Copyright (c) DGP Studio. All rights reserved.
 // Licensed under the MIT license.
 
+using Microsoft.UI.Xaml.Controls;
 using Snap.Hutao.Factory.ContentDialog;
 using Snap.Hutao.Model.Entity;
 using Snap.Hutao.Service.Cultivation;
@@ -110,10 +111,17 @@ internal sealed partial class CultivationViewModel : Abstraction.ViewModel
             return;
         }
 
-        await cultivationService.RemoveProjectAsync(project).ConfigureAwait(false);
-        await taskContext.SwitchToMainThreadAsync();
-        ArgumentNullException.ThrowIfNull(Projects);
-        SelectedProject = Projects.FirstOrDefault();
+        ContentDialogResult result = await contentDialogFactory
+            .CreateForConfirmCancelAsync(SH.ViewModelCultivationRemoveProjectTitle, SH.ViewModelCultivationRemoveProjectContent)
+            .ConfigureAwait(false);
+
+        if (result is ContentDialogResult.Primary)
+        {
+            await cultivationService.RemoveProjectAsync(project).ConfigureAwait(false);
+            await taskContext.SwitchToMainThreadAsync();
+            ArgumentNullException.ThrowIfNull(Projects);
+            SelectedProject = Projects.FirstOrDefault();
+        }
     }
 
     private async ValueTask UpdateEntryCollectionAsync(CultivateProject? project)
